mcpserver1/pyproject.toml
[project]
name = "mcpserver1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.8.0",
]
